debug fix

+++ src/main/java/com/example/slstore/customer/dto/PercentageCouponDto.java
@@ -3,19 +3,19 @@ package com.example.slstore.customer.dto;
 import java.time.LocalDate;
 
 public class PercentageCouponDto extends CouponDto {
-    private double discountPercentage;
+    private int discountPercentage;
 
     public PercentageCouponDto(Integer id, String code, String description,
-            LocalDate validFrom, LocalDate validUntil, double discountPercentage) {
+            LocalDate validFrom, LocalDate validUntil, int discountPercentage) {
         super(id, code, description, validFrom, validUntil);
         this.discountPercentage = discountPercentage;
     }
 
-    public double getDiscountPercentage() {
+    public int getDiscountPercentage() {
         return discountPercentage;
     }
 
-    public void setDiscountPercentage(double discountPercentage) {
+    public void setDiscountPercentage(int discountPercentage) {
         this.discountPercentage = discountPercentage;
     }
